+++ apollo-federation-types/src/lib.rs
@@ -1,10 +1,12 @@
 #[cfg(feature = "build")]
-pub mod build;
+pub mod rover;
 
 #[cfg(feature = "build_plugin")]
 pub mod build_plugin;
 
 #[cfg(feature = "config")]
 pub mod config;
+
+pub mod javascript;
 
 pub(crate) type UncaughtJson = std::collections::BTreeMap<String, serde_json::Value>;