+++ apollo-federation-types/src/rover/mod.rs
@@ -1,14 +1,14 @@
+//! This module contains the interface between Rover and its `supergraph` binaries.
+
 pub use error::{BuildError, BuildErrorType, BuildErrors};
 pub use hint::BuildHint;
 pub use output::BuildOutput;
-pub use subgraph_definition::SubgraphDefinition;
 
 use crate::build_plugin::{BuildMessageLevel, PluginFailureReason, PluginResult};
 
 mod error;
 mod hint;
 mod output;
-mod subgraph_definition;
 
 /// The type representing the result of a supergraph build (for any version)
 pub type BuildResult = Result<BuildOutput, BuildErrors>;